third_party/blink/web_tests/virtual/portals/external/wpt/portals/idlharness.window-expected.txt

This is a testharness.js-based test.
Found 5 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Stringification of window.portalHost
  assert_class_string: class string of window.portalHost expected "[object PortalHost]" but got "[object Null]"
[FAIL] PortalHost interface: window.portalHost must inherit property "postMessage(any, optional StructuredSerializeOptions)" with the proper type
  assert_inherits: provided value is not an object
[FAIL] PortalHost interface: calling postMessage(any, optional StructuredSerializeOptions) on window.portalHost with too few arguments must throw TypeError
  assert_inherits: provided value is not an object
[FAIL] PortalHost interface: window.portalHost must inherit property "onmessage" with the proper type
  assert_inherits: provided value is not an object
[FAIL] PortalHost interface: window.portalHost must inherit property "onmessageerror" with the proper type
  assert_inherits: provided value is not an object
Harness: the test ran to completion.
